+++ node/resolvers/directives/withPermissions.ts
@@ -7,11 +7,10 @@ import { SchemaDirectiveVisitor } from 'graphql-tools'
 import type StorefrontPermissions from '../../clients/storefrontPermissions'
 
 export const getUserPermission = async (
-  storefrontPermissions: StorefrontPermissions
+  storefrontPermissions: StorefrontPermissions,
+  app = 'vtex.b2b-organizations@0.x'
 ) => {
-  const result = await storefrontPermissions.checkUserPermission(
-    'vtex.b2b-organizations@0.x'
-  )
+  const result = await storefrontPermissions.checkUserPermission(app)
 
   return result?.data?.checkUserPermission ?? null
 }
@@ -20,21 +19,22 @@ export class WithPermissions extends SchemaDirectiveVisitor {
   public visitFieldDefinition(field: GraphQLField<any, any>) {
     const { resolve = defaultFieldResolver } = field
 
-    field.resolve = async (
-      root: any,
-      args: any,
-      context: Context,
-      info: any
-    ) => {
+    field.resolve = async (root: any, args: any, context: any, info: any) => {
       const {
         clients: { storefrontPermissions },
         vtex: { adminUserAuthToken, logger },
       } = context
 
       const appClients = context.vtex as any
+      const sender =
+        context?.graphql?.query?.senderApp ??
+        context?.graphql?.query?.extensions?.persistedQuery?.sender ??
+        context?.request?.header['x-b2b-senderapp'] ??
+        undefined
 
       appClients.storefrontPermissions = await getUserPermission(
-        storefrontPermissions
+        storefrontPermissions,
+        sender
       ).catch((error: any) => {
         if (!adminUserAuthToken) {
           logger.error({