+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Bug fix on `withPermissions` directive to allow the use of the `sender` directive
+
 ## [0.63.3] - 2025-01-28
 ### Fixed
 - Fix return from `removeUserWithEmail`